refactor API classes
