+++ frontend/src/services/search/searchfetcher/APISearchFetcher.ts
@@ -21,9 +21,7 @@ export class APISearchFetcher extends SearchFetcher {
       // await new Promise((resolve) => setTimeout(resolve, 13250));
 
       const response: SearchAPIResponse =
-        (await this.searchApi.searchOpportunities(
-          searchInputs,
-        )) as SearchAPIResponse;
+        await this.searchApi.searchOpportunities(searchInputs);
       response.actionType = searchInputs.actionType;
       response.fieldChanged = searchInputs.fieldChanged;
